requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
idna==2.8
numpy==1.18.0
Pillow==6.2.1
pycparser==2.19
python-telegram-bot==12.2.0
requests==2.22.0
six==1.13.0
tornado==6.0.3
tqdm==4.41.0
urllib3==1.25.7
